fix:初始化项目

+++ rbm/src/main/java/com/blog/rbm/common/Constants.java
@@ -5,7 +5,7 @@ public class Constants {
     public static String DEFAULT_LOG_TYPE = "1";
 
     // redis的过期时间
-    public static int USER_TOKEN_EXPIRE = 60*60*2;
+    public static long USER_TOKEN_EXPIRE = 60*60*2;
 
     public static String SUCCESS_CODE = "200";
 }
